requirements.txt
beautifulsoup4
python-telegram-bot